+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Renamed `PollData` and `PollAnswer` to `Poll` and `ProvidedAnswer` ([#536]((https://github.com/desmos-labs/desmos/issues/536)))
 - Enabled snapshot by default ([#529](https://github.com/desmos-labs/desmos/pull/529))
 - Improved `x/profiles` params structure [#538](https://github.com/desmos-labs/desmos/issues/538)
+- Made oracle-related data on-chain parameters ([#547](https://github.com/desmos-labs/desmos/pull/547))
 
 ## Version 0.17.0
 ### Changes